This is Jixie's fir commit into jixie's branch       modified:   gen_evt.f       modified:   guout.f       modified:   gustep.f       deleted:    test I cleaned some comments in gen_evt.f, also fixed several bugs in it
In guout.f I comment out a if statement which is requiring there
are hits in calorimeter in order to write an event into the ntuple.
This is temperaly change. I am trying to write all event into the
ntuple so I can do charge normalization. I will change it back
once I finish my job.

gustep.f, I added a comment. I verified that call step_tracker()
will not affect the speed.  Still need to find out why this program runs
so slow.

I remove test since it is not used any where and donot know what it is.